Defense Sequence Manager
I think I, at least for now, have managed to get the Defense Sequence Manager in a state that I can move on from. Next up is the area of effect entry, controller and manager. After that will be the status effect entry, container and manager and then I can finally hook in the entity scripts. The main goal after doing all of this will to be to go through, test things and make sure all of the functionality for the more complex data is attached fully and then to make sure it is correct. I wont have it balanced but as long as it is working, I will worry about balance later.

The main goal is to convert the rest of my code from the previous version I was using to learn data structure, Unities built in functionality and other functionalities. I want to learn to be both structurally refined and be up to the modern standard for things related to coding and other fields of game development.